Delete gallery model

+++ packages/Order/src/Database/Migrations/2019_10_31_173337_create_order_products_table.php
@@ -20,10 +20,10 @@ class CreateOrderProductsTable extends Migration
 
             $table->decimal('total', 12,4)->default(0);
 
-            $table->integer('order_id')->unsigned()->nullable();
+            $table->bigInteger('order_id')->unsigned()->nullable();
             $table->foreign('order_id')->references('id')
                   ->on('orders')->onUpdate('cascade')->onDelete('set null');
-            $table->integer('product_id')->unsigned()->nullable();
+            $table->bigInteger('product_id')->unsigned()->nullable();
             $table->foreign('product_id')->references('id')
                 ->on('products')->onUpdate('cascade')->onDelete('set null');
